+++ .gitignore
@@ -1,5 +1,8 @@
 # Generated files
 *.mocks.dart
+*.g.dart
+android/**/generated/
+ios/**/Generated/
 
 # Miscellaneous
 *.class
