+++ ios/Classes/InstabugApiImpl.h
@@ -1,0 +1,2 @@
+@interface InstabugApiImpl : NSObject <InstabugApi>
+@end
