+++ pigeons/instabug.api.dart
@@ -1,0 +1,6 @@
+import 'package:pigeon/pigeon.dart';
+
+@HostApi()
+abstract class InstabugApi {
+  void start(String token, List<String> invocationEvents);
+}
